Add rudimentary health targets
Add a `health` make target to the stacks that outputs a summary of the
stacks health to quickly identify whether it is running. These targets
just format the output of the docker health information and therefore
work with any docker compose stack.

The output of the curl or wget commands can wrap down multiple lines and
make the output messy. While work could be done to truncate these based
on the terminal width, this didn't feel essential so instead the curl
and wget outputs have been made to be less verbose to limit the
ugliness.