Add a target to clone all services for stack
Add a make target that clones all of the git repos for the services
defined in the docker compose stack.

This change also standardises the path override for the repo location,
replacing the individual `ROOT_${service_name}` with a single
`DP_REPO_DIR` env var. While this change does reduce the
customisability, it will reduce the issues we see when the repos are not
cloned beside each other. The default behaviour will continue to be the
same with the app repos being cloned in the same directory as
dp-compose.

The documentation needs updating, but this will come in a subsequent
change.