Remove unused redis instance
Remove the redis instance from the original docker compose stack as it
is no longer needed. This was introduced as part of a piece of work
exploring potential options to move the sessions logic out of zebedee.
This work was never taken forwards past the initial investigations so
this redis instance was never used.