Add a v1 compatible v2 stack for transition
Provide a transitional stack in v2 that provides a drop in replacement
for the original `docker-compose.yml` found in the root of the repo.

The make commands are all aligned to v2 so working with the stack will
be slightly different, but the stack is provided to enable a smooth
transition to v2.