Add READMEs for stacks missing them
Make sure that all stacks have basic READMEs and remove all of the stack
specific information from the general stacks README.

Reduce the duplication in the stacks README and generally tidy it up a
bit. More work is needed as the make targets listed are incompleted, but
this can come in a separate change.